Changed conception of converting to tensor.
Previously used trait ConvStateToTensor opertaing on infoset and
converting it to tensor.
Going to be replaced with WayToTensor which does not converts InfoSet,
but is type holder for ConvertToTensor trait which can be implemented
for types to be converted to tensor with regard to WayToTensor type.
This holds generic demand and is easier to contraint in implementations